nerox-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.2
python-dotenv>=1.0.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.10.0
email-validator>=2.1.1
pydantic-settings>=2.3.0
python-multipart>=0.0.9

# ── Phase 3: AI Fingerprinting & Detection Engine ──────────────────────────
# Install with: pip install torch torchvision opencv-python faiss-cpu numpy
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# ── Phase 5: Invisible Watermarking + Ownership Trace ──────────────────────
# Uses existing deps: opencv-python (DCT transforms), numpy, hashlib/hmac (stdlib)
# No additional packages required.
